requirements.txt
asgiref==3.8.1
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
mypy==1.10.1
mypy-extensions==1.0.0
mysqlclient==2.2.4
packaging==24.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
